+++ src/main/java/com/finanzas/breadcredit/entity/ProductPurchaseId.java
@@ -15,8 +15,8 @@ import java.io.Serializable;
 public class ProductPurchaseId implements Serializable {
 
     @Column(name = "purchase_id", nullable = false)
-    private Integer purchaseId;
+    private Long purchaseId;
     @Column(name = "product_id", nullable = false)
-    private Integer productId;
+    private Long productId;
 
 }
