+++ src/main/java/com/finanzas/breadcredit/repository/ProductRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-    Optional<List<Product>> findByAdmin_Id(Integer adminId);
+public interface ProductRepository extends JpaRepository<Product, Long> {
+    Optional<List<Product>> findByAdmin_Id(Long adminId);
 }
