+++ src/main/java/com/finanzas/breadcredit/repository/PurchaseRepository.java
@@ -3,5 +3,5 @@ package com.finanzas.breadcredit.repository;
 import com.finanzas.breadcredit.entity.Purchase;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PurchaseRepository extends JpaRepository<Purchase, Integer> {
+public interface PurchaseRepository extends JpaRepository<Purchase, Long> {
 }
